+++ composer/autoload_classmap.php
@@ -2302,6 +2302,8 @@ return array(
     'Peast\\Syntax\\ES2023\\Features' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/ES2023/Features.php',
     'Peast\\Syntax\\ES2024\\Features' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/ES2024/Features.php',
     'Peast\\Syntax\\ES2025\\Features' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/ES2025/Features.php',
+    'Peast\\Syntax\\ES2026\\Features' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/ES2026/Features.php',
+    'Peast\\Syntax\\ES2027\\Features' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/ES2027/Features.php',
     'Peast\\Syntax\\EncodingException' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/EncodingException.php',
     'Peast\\Syntax\\EventsEmitter' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/EventsEmitter.php',
     'Peast\\Syntax\\Exception' => $vendorDir . '/mck89/peast/lib/Peast/Syntax/Exception.php',