+++ mck89/peast/lib/Peast/Peast.php
@@ -116,6 +116,22 @@ namespace Peast;
  * Returns a parser instance with ES2025 features for the given source. See Peast
  * class documentation to understand function arguments.
  *
+ * @method static Syntax\Parser ES2026(string $source, array $options = array())
+ * Returns a parser instance with ES2026 features for the given source. See Peast
+ * class documentation to understand function arguments.
+ *
+ * @method static Syntax\Parser ES17(string $source, array $options = array())
+ * Returns a parser instance with ES2026 features for the given source. See Peast
+ * class documentation to understand function arguments.
+ *
+ * @method static Syntax\Parser ES2027(string $source, array $options = array())
+ * Returns a parser instance with ES2027 features for the given source. See Peast
+ * class documentation to understand function arguments.
+ *
+ * @method static Syntax\Parser ES18(string $source, array $options = array())
+ * Returns a parser instance with ES2027 features for the given source. See Peast
+ * class documentation to understand function arguments.
+ *
  * @method static Syntax\Parser latest(string $source, array $options = array())
  * Returns an instance of the latest parser version for the given source. See
  * Peast class documentation to understand function arguments.
@@ -153,7 +169,9 @@ class Peast
         "ES13" => "ES2022",
         "ES14" => "ES2023",
         "ES15" => "ES2024",
-        "ES16" => "ES2025"
+        "ES16" => "ES2025",
+        "ES17" => "ES2026",
+        "ES18" => "ES2027"
     );
     
     /**