+++ mck89/peast/lib/Peast/Syntax/ES2027/Features.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * This file is part of the Peast package
+ *
+ * (c) Marco Marchiò <marco.mm89@gmail.com>
+ *
+ * For the full copyright and license information refer to the LICENSE file
+ * distributed with this source code
+ */
+namespace Peast\Syntax\ES2027;
+
+/**
+ * ES2027 features class
+ * 
+ * @author Marco Marchiò <marco.mm89@gmail.com>
+ * 
+ * @codeCoverageIgnore
+ */
+class Features extends \Peast\Syntax\ES2026\Features
+{
+    /**
+     * Explicit resource management
+     *
+     * @var bool
+     */
+    public $explicitResourceManagement = true;
+}