+++ mck89/peast/doc/changelog.md
@@ -1,6 +1,16 @@
 Changelog
 ==========
 
+#### 1.18.0
+* Added ES2026 parser with no additional features
+* Added ES2027 parser and implemented explicit resource management
+
+#### 1.17.7
+* Fixed deprecation warning about null array index on PHP 8.5 when parsing trailing comments
+
+#### 1.17.6
+* Fixed bug where single BlockStatement where wrongly expanded removing brackets
+
 #### 1.17.5
 * Fixed bug when parsing SpreadElement followed by NewExpression between parenthesis